pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kiosk-control"
version = "0.0.2"
description = "Policy-driven Raspberry Pi kiosk controller (Chromium CDP + plugins + D-Bus) with optional touch overlay UI."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "kiosk-control contributors"}]
keywords = ["raspberry-pi", "kiosk", "chromium", "cdp", "nightscout", "home-assistant", "dbus"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
]

dependencies = [
  "PyYAML>=6.0",
  "aiohttp>=3.9",
  "websockets>=12.0",
  "dbus-next>=0.2.3",
]

[project.optional-dependencies]
nightscout = [
  "python-socketio>=5.11.0",
]
homeassistant = []
input = ["evdev>=1.7.1"]
ui = []

dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.3",
]

[project.scripts]
kiosk-control = "kiosk_control.cli:main"
kiosk-overlay = "kiosk_overlay.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
  "E", "F", "W",
  "I",
  "UP",
  "B",
  "SIM",
]
ignore = [
  "E501",  # handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.ruff.lint.per-file-ignores]
"src/kiosk_control/dbus_service.py" = ["F821", "UP037"]
